general login user-flow achieved
