+++ components/Navbar/Navbar.scss
@@ -15,3 +15,8 @@ a {
   color: $green;
   padding: 5px;
 }
+
+a > img {
+  max-width: 50px;
+  margin-right: auto;
+}
