+++ server/config/passport-config.js
@@ -13,13 +13,13 @@ module.exports = {
     }, (email, password, done) => {
       User.findOne({where: {email}})
       .then((user) => {
-        console.log('{PASSPORT-CONFIG} USER.EMAIL: ', user.email);
+        // console.log('{PASSPORT-CONFIG} USER.EMAIL: ', user.email);
         if(!user || !authHelper.comparePass(password, user.password)) {
-          console.log('{PASSPORT-CONFIG} !USER OR !AUTHHELPER.COMPAREPASS', !user, !authHelper.comparePass(password, user.password));
+          // console.log('{PASSPORT-CONFIG} !USER OR !AUTHHELPER.COMPAREPASS', !user, !authHelper.comparePass(password, user.password));
           // console.log('USER.email: ', user.email);
           // console.log('PASSWORD FROM CLIENT: ', password);
           // console.log('FROM DBUSER.PASSWORD: ', user.password);
-          // return done(null, false, {message: 'Invalid email or password'});
+          return done(null, false, {message: 'Invalid email or password'});
         }
         // this puts user object in req.session.passport.user = {id: ...}
         console.log('{PASSPORT-CONFIG} USER AUTHENTICATED')
